# data_object_service.swagger.json
{
  "swagger": "2.0",
  "basePath": "api/ga4gh/dos/v1",
  "info": {
    "title": "Data Object Service",
    "version": "0.2.0"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/dataobjects/list": {
      "post": {
        "summary": "List the Data Objects",
        "operationId": "ListDataObjects",
        "responses": {
          "200": {
            "description": "The Data Objects were listed successfully.",
            "schema": {
              "$ref": "#/definitions/ListDataObjectsResponse"
            }
          },
          "400": {
            "description": "The request is malformed.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "The request is unauthorized.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "The requester is not authorized to perform this action.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ListDataObjectsRequest"
            }
          }
        ],
        "tags": [
          "DataObjectService"
        ],
        "x-swagger-router-controller": "ga4gh.dos.server"
      }
    },
...
  },
  "definitions": {
    "SystemMetadata": {
      "type": "object",
      "additionalProperties": true,
      "description": "OPTIONAL\nThese values are reported by the underlying object store.\nA set of key-value pairs that represent system metadata about the object."
    },
    "UserMetadata": {
      "type": "object",
      "additionalProperties": true,
      "description": "OPTIONAL\nA set of key-value pairs that represent metadata provided by the uploader."
    },
    "Checksum": {
      "type": "object",
      "properties": {
        "checksum": {
          "type": "string",
          "description": "REQUIRED\nThe hex-string encoded checksum for the Data."
        },
        "type": {
          "type": "string",
          "description": "OPTIONAL\nThe digest method used to create the checksum. If left unspecified md5\nwill be assumed.\n\npossible values:\nmd5                # most blob stores provide a checksum using this\nmultipart-md5      # multipart uploads provide a specialized tag in S3\nsha256\nsha512"
        }
      }
    },
    "ChecksumRequest": {
      "type": "object",
      "properties": {
        "checksum": {
          "type": "string",
          "description": "REQUIRED\nThe hexlified checksum that one would like to match on."
        },
        "type": {
          "type": "string",
          "description": "OPTIONAL\nIf provided will restrict responses to those that match the provided\ntype.\n\npossible values:\nmd5                # most blob stores provide a checksum using this\nmultipart-md5      # multipart uploads provide a specialized tag in S3\nsha256\nsha512"
        }
      }
    },
...
    "DataObject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "REQUIRED\nAn identifier unique to this Data Object."
        },
        "name": {
          "type": "string",
          "description": "OPTIONAL\nA string that can be optionally used to name a Data Object."
        },
        "size": {
          "type": "string",
          "format": "int64",
          "description": "REQUIRED\nThe computed size in bytes."
        },
        "created": {
          "type": "string",
          "format": "date-time",
          "description": "REQUIRED\nTimestamp of object creation in RFC3339."
        },
        "updated": {
          "type": "string",
          "format": "date-time",
          "description": "OPTIONAL\nTimestamp of update in RFC3339, identical to create timestamp in systems\nthat do not support updates."
        },
        "version": {
          "type": "string",
          "description": "OPTIONAL\nA string representing a version."
        },
        "mime_type": {
          "type": "string",
          "description": "OPTIONAL\nA string providing the mime-type of the Data Object.\nFor example, \"application/json\"."
        },
        "checksums": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Checksum"
          },
          "description": "REQUIRED\nThe checksum of the Data Object. At least one checksum must be provided."
        },
        "urls": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/URL"
          },
          "description": "OPTIONAL\nThe list of URLs that can be used to access the Data Object."
        },
        "description": {
          "type": "string",
          "description": "OPTIONAL\nA human readable description of the contents of the Data Object."
        },
        "aliases": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "OPTIONAL\nA list of strings that can be used to find this Data Object.\nThese aliases can be used to represent the Data Object's location in\na directory (e.g. \"bucket/folder/file.name\") to make Data Objects\nmore discoverable. They might also be used to represent"
        }
      }
    },
...
    "ListDataObjectsRequest": {
      "type": "object",
      "properties": {
        "alias": {
          "type": "string",
          "description": "OPTIONAL\nIf provided will only return Data Objects with the given alias."
        },
        "url": {
          "type": "string",
          "description": "OPTIONAL\nIf provided will return only Data Objects with a that URL matches\nthis string."
        },
        "checksum": {
          "$ref": "#/definitions/ChecksumRequest",
          "description": "OPTIONAL\nIf provided will only return data object messages with the provided\nchecksum. If the checksum type is provided"
        },
        "page_size": {
          "type": "integer",
          "format": "int32",
          "description": "OPTIONAL\nSpecifies the maximum number of results to return in a single page.\nIf unspecified, a system default will be used."
        },
        "page_token": {
          "type": "string",
          "description": "OPTIONAL\nThe continuation token, which is used to page through large result sets.\nTo get the next page of results, set this parameter to the value of\n`next_page_token` from the previous response."
        }
      },
      "description": "Allows a requester to list and filter Data Objects. Only Data Objects\nmatching all of the requested parameters will be returned."
    },
    "ListDataObjectsResponse": {
      "type": "object",
      "properties": {
        "data_objects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataObject"
          },
          "description": "The list of Data Objects."
        },
        "next_page_token": {
          "type": "string",
          "description": "The continuation token, which is used to page through large result sets.\nProvide this value in a subsequent request to return the next page of\nresults. This field will be empty if there aren't any additional results."
        }
      },
      "description": "A list of Data Objects matching the requested parameters, and a paging\ntoken, that can be used to retrieve more results."
    },
    "URL": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string",
          "description": "REQUIRED\nA URL that can be used to access the file."
        },
        "system_metadata": {
          "$ref": "#/definitions/SystemMetadata"
        },
        "user_metadata": {
          "$ref": "#/definitions/UserMetadata"
        }
      }
    },
...
    "ErrorResponse": {
      "description": "An object that can optionally include information about the error.",
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "description": "A detailed error message."
        },
        "status_code": {
          "type": "integer",
          "description": "The integer representing the HTTP status code (e.g. 200, 404)."
        }
      }
    }
  }
}